cleaning up branch before merge to main

+++ GreenMile/src/Web/Pages/FoodTracker/Index.cshtml.cs
@@ -3,6 +3,10 @@ using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 
+using Newtonsoft.Json;
+
+using Web.Data;
+using Web.Lib;
 using Web.Models;
 using Web.Services;
 
@@ -13,41 +17,91 @@ namespace Web.Pages.FoodTracker
     {
         private readonly FoodItemService _fooditemService;
         private readonly UserManager<User> _userManager;
+
+
+        private readonly FoodCategoryService _foodcategoryService;
+
         private readonly IHouseholdService _householdService;
+        private readonly IHttpContextAccessor _contextAccessor;
 
-        [BindProperty]
-        public int Id { get; set; }
-        public int Count { get; set; }
-        public IEnumerable<FoodItem> FoodItemList { get; set; }
+       
 
 
-        public IndexModel(FoodItemService fooditemService, UserManager<User> userManager, IHouseholdService householdService)
+
+        public IndexModel(FoodItemService fooditemService, IHouseholdService householdService, IHttpContextAccessor contextAccessor, FoodCategoryService foodcategoryService, UserManager<User> userManager)
+
         {
             _fooditemService = fooditemService;
             _userManager = userManager;
             _householdService = householdService;
+
+            _contextAccessor = contextAccessor;
+
+            _foodcategoryService = foodcategoryService;
+
+
         }
+        public List<FoodItem> FoodItemList { get; set; }
+
+        public IEnumerable<Category> Categories { get; set; }
+
+        [BindProperty]
+        public int Id { get; set; }
+
+        [BindProperty]
+        public string action_type { get; set; }
+
+       
+
+        
+
+        public int Count { get; set; }
+        public string Name { get; set; }
 
         public async Task OnGetAsync()
         {
             var user = await _userManager.GetUserAsync(HttpContext.User);
             var household = (await _householdService.RetrieveHouseholdDetails(user.HouseholdId ?? -1)).Value;
             FoodItemList = _fooditemService.GetAll(household);
+            Name = household.Name;
             Count = FoodItemList.Count();
-        }
 
+
+            var newcategory = new Category()
+            {
+                Name = "Fruit",
+                Description = "Fruits are healthy"
+
+            };
+            _foodcategoryService.AddCategory(newcategory);
+
+        }
+       
         public async Task<IActionResult> OnPostAsync()
         {
+
             FoodItem? food = await _fooditemService.GetFoodItemById(Id);
+
             if (food != null)
             {
-                _fooditemService.DeleteFoodItem(food);
-                return Redirect("/Index");
+                if(action_type == "delete")
+                {
+                    _fooditemService.DeleteFoodItem(food);
+                    return Redirect("/FoodTracker");
+                }
+                
             }
             else
             {
                 return Page();
             }
+
+            return Redirect("/FoodTracker");
+
+ 
+       
+
+
         }
     }
 }
